make all the requested changes from PR#2
