+++ config/urls.py
@@ -2,7 +2,6 @@ from django.conf import settings
 from django.urls import include, path, re_path
 from django.conf.urls.static import static
 from django.contrib import admin
-# from django.views.decorators.csrf import csrf_exempt
 from django.views.generic import TemplateView
 from django.views import defaults as default_views
 from rest_framework.authtoken.views import obtain_auth_token
@@ -40,7 +39,6 @@ urlpatterns += [
 # API URLS
 urlpatterns += [
     # GraphQL
-    # path("graphql/", csrf_exempt(FileUploadGraphQLView.as_view(graphiql=True, pretty=True))),
 ]
 
 if settings.DEBUG:
